requirements-web.txt
flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
tensorflow==2.20.0
numpy>=1.24.0
Pillow>=10.2.0
